Update README.md
i. An array is a variable that can store multiple values.
ii. Arrays have 0 as the first index.
iii. To create an array, define the data type and specify the name of the array followed by square brackets [].
iv. To insert values to it, use a comma-separated list, inside curly braces:
      int myNumbers[] = {25, 50, 75, 100};
v. It has contiguous memory.